requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
enum34==1.1.10
exif==0.8.2
ExifRead==2.1.2
gpsphoto==2.2.3
idna==2.9
numpy==1.18.2
opencv-python==4.2.0.34
piexif==1.1.3
Pillow==7.1.1
pycparser==2.20
pyzbar==0.1.8
requests==2.23.0
six==1.14.0
urllib3==1.25.8
